Fix resolving archful gems with rails-ex app.
The Bundler version 2.3.26 does not correctly resolve archful
packages that contain Ruby version constraints.

For bundler versions >= 2, an option can be set to install only
the "ruby" platform gems. These are source-only libraries that compile
their extensions locally. Therefore this kind of workaround comes with
the burden of having to ensure that all the build dependencies that the
gems require are present during build.